Split continuations.js into two files and add build step to assemble them
