+++ Makefile
@@ -24,7 +24,7 @@ all: check \
 	dist/module_webworker_dev.js
 	echo -e "\nSUCCESS!"
 
-src/core/pyodide_pre.o: src/js/_pyodide.out.js src/core/pre.js src/core/continuations.js
+src/core/pyodide_pre.o: src/js/_pyodide.out.js src/core/pre.js src/core/create_invokes.out.js
 # Our goal here is to inject src/js/_pyodide.out.js into an archive file so that
 # when linked, Emscripten will include it. We use the same pathway that EM_JS
 # uses, but EM_JS is itself unsuitable. Why? Because the C preprocessor /
@@ -56,7 +56,7 @@ src/core/pyodide_pre.o: src/js/_pyodide.out.js src/core/pre.js src/core/continua
 	echo '()<::>{' >> tmp.dat             # zero argument argspec and start body
 	cat src/js/_pyodide.out.js >> tmp.dat # All of _pyodide.out.js is body
 	echo '}' >> tmp.dat                   # Close function body
-	cat src/core/continuations.js >> tmp.dat # Execute continuations.js
+	cat src/core/create_invokes.out.js >> tmp.dat
 	cat src/core/pre.js >> tmp.dat           # And pre.js
 	echo "pyodide_js_init();" >> tmp.dat     # Then execute the function.
 
@@ -127,6 +127,9 @@ node_modules/.installed : src/js/package.json src/js/package-lock.json
 
 dist/pyodide.js src/js/_pyodide.out.js: src/js/*.ts src/js/pyproxy.gen.ts src/js/error_handling.gen.ts node_modules/.installed
 	cd src/js && npm run tsc && node esbuild.config.mjs && cd -
+
+src/core/create_invokes.out.js : src/core/create_invokes.mjs src/core/runtime_wasm.mjs
+	cd src/core && npx esbuild --bundle create_invokes.mjs --outfile=create_invokes.out.js
 
 dist/package.json : src/js/package.json
 	cp $< $@
