+++ src/js/test/unit/continuations.test.mjs
@@ -1,10 +1,19 @@
-globalThis.Module = {};
 import { expect } from "chai";
 
 import { readFileSync, writeFileSync } from "node:fs";
 import * as vm from "node:vm";
 import loadWabt from "wabt";
 import { URL } from "url"; // in Browser, the URL in native accessible on window
+import {
+  WasmModule,
+  CodeSection,
+  ImportSection,
+  emscriptenSigToWasm,
+  TypeSection,
+  WASM_PRELUDE,
+  insertSectionPrefix,
+} from "../../../core/runtime_wasm.mjs";
+import { createInvokeModule } from "../../../core/create_invokes.mjs";
 
 const __dirname = new URL(".", import.meta.url).pathname;
 
@@ -80,9 +89,9 @@ function findSection(mod, section) {
   }
 }
 
-const source = readFileSync("../core/continuations.js", { encoding: "utf8" });
-const script = new vm.Script(source);
-script.runInThisContext();
+// const source = readFileSync("../core/continuations.js", { encoding: "utf8" });
+// const script = new vm.Script(source);
+// script.runInThisContext();
 
 // Monkey patch to prevent it from creating an actual WebAssembly.Module
 // as the result so we can assert on the generated bytes.
